+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- New property added to entity:
+
+  | Entity         | Properties         |
+  | -------------- | ------------------ |
+  | `snyk_finding` | `originalSeverity` |
+
 ## [2.2.1] - 2022-07-19
 
 ### Fixed